Move viz backend setup variable and function to mne/viz/utils
