+++ mne/viz/backend.py
@@ -1,5 +1,5 @@
 import sys
-from .. import _BACKEND
+from .utils import _BACKEND
 
 sys.stderr.write('Using {} backend.\n'.format(_BACKEND))
 
@@ -7,5 +7,5 @@ if _BACKEND == 'mlab':
     from .mlab_backend import *
 else:
     import warnings
-    warnings.warn('_BACKEND should be "mlab". '
-'{} was given.'.format(_BACKEND))
+    warnings.warn('_BACKEND should be "mlab" : '
+                  '{} was given.'.format(_BACKEND))
