+++ mne/viz/utils.py
@@ -15,6 +15,8 @@ from functools import partial
 import difflib
 import webbrowser
 import tempfile
+import os
+import importlib
 import numpy as np
 from copy import deepcopy
 from distutils.version import LooseVersion
@@ -42,6 +44,46 @@ _channel_type_prettyprint = {'eeg': "EEG channel", 'grad': "Gradiometer",
                              'eog': "EOG channel", 'ecg': "ECG sensor",
                              'emg': "EMG sensor", 'ecog': "ECoG channel",
                              'misc': "miscellaneous sensor"}
+
+# Set the default backend
+default_backend = 'mlab'
+
+
+global _BACKEND
+try:
+    if _BACKEND is None:
+        _BACKEND = os.environ.get('MNE_3D_BACKEND', default_backend)
+except NameError:
+    _BACKEND = os.environ.get('MNE_3D_BACKEND', default_backend)
+
+
+def set_backend(backend_name):
+    """Set the backend for MNE.
+
+    The backend will be set as specified and operations will use
+    that backend
+
+    Parameters
+    ----------
+    backend_name : {'mlab'}, default is 'mlab'
+    """
+    global _BACKEND
+    _BACKEND = backend_name
+    from . import backend
+    importlib.reload(backend)
+
+
+def get_backend():
+    """Return the backend currently used.
+
+    Returns
+    -------
+    backend_used : str
+        the backend currently in use
+    """
+    global _BACKEND
+    backend_used = _BACKEND
+    return backend_used
 
 
 def _setup_vmin_vmax(data, vmin, vmax, norm=False):
